backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# psycopg[binary]==3.3.3
pydantic>=2.9.0
pydantic-settings>=2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-multipart==0.0.6
pandas>=2.3.3
alembic==1.13.0
pytest==7.4.3
httpx==0.25.2
